pipectl: don't unlink pipe if pipe has been unlinked by another instance
This is done by checking fstat()'s st_nlink count. If it's 0, the file
has been unlinked by another instance and there is no reason to remove
it.

This doesn't protect against someone hardlinking the pipe to somewhere
else before removing the original, but that doesn't happen in normal use
and is basically impossible to check for.

Fixes #2.